add service to reload auffaelligkeiten
