+++ frontend/src/api/service/AdministrationService.ts
@@ -1,12 +1,17 @@
+import type ReloadAuffaelligkeitenDTO from "@/types/config/ReloadAuffaelligkeitenDTO";
+
 import FetchService from "@/api/service/FetchService";
 
 export default class AdministrationService {
   private static readonly ENDPOINT: string =
     "api/dave-backend-service/administration";
 
-  static reloadUnauffaelligerTag(datum: Date): Promise<void> {
-    return FetchService.getData(
-      `${this.ENDPOINT}/reload-unauffaelliger-tag?date_to_reload=${datum}`,
+  static reloadUnauffaelligerTag(
+    data: ReloadAuffaelligkeitenDTO
+  ): Promise<void> {
+    return FetchService.postData(
+      data,
+      `${this.ENDPOINT}/reload-unauffaelliger-tag`,
       "Beim Neuladen des Unauffälligen Tages ist ein Fehler aufgetreten."
     );
   }
