+++ frontend/src/components/administration/ReloadAuffaelligkeitenCard.vue
@@ -1,0 +1,150 @@
+<template>
+  <v-card
+    variant="outlined"
+    class="ml-3 mt-3"
+    :max-width="maxWidth"
+  >
+    <v-card-title>
+      <v-row
+        dense
+        no-gutters
+      >
+        <span>Tag auf Auffälligkeiten prüfen</span>
+        <v-spacer />
+        <v-icon
+          v-tooltip:end="TOOLTIP_RELOAD_UNAUFFAELLIGER_TAG"
+          size="small"
+          icon="mdi-information-outline"
+        />
+      </v-row>
+    </v-card-title>
+    <v-card-text>
+      <v-date-picker v-model="choosenDate">
+        <template #title />
+      </v-date-picker>
+    </v-card-text>
+    <v-card-actions>
+      <v-spacer />
+      <v-btn
+        variant="outlined"
+        text="Prüfen"
+        @click="reloadUnauffaelligerTag"
+      />
+    </v-card-actions>
+  </v-card>
+  <v-dialog
+    v-model="openDialogModel"
+    persistent
+    width="800"
+  >
+    <v-card>
+      <v-card-title>
+        <v-icon
+          start
+          icon="mdi-calendar-search-outline"
+        />
+        Tag auf Auffälligkeiten prüfen
+      </v-card-title>
+      <v-card-text>
+        <span v-html="dialogtext" />
+      </v-card-text>
+      <v-card-actions>
+        <v-spacer />
+        <v-btn
+          color="red-lighten-1"
+          text="Ja"
+          variant="elevated"
+          @click="confirmReload"
+        />
+        <v-btn
+          color="tertiary"
+          text="Nein"
+          variant="elevated"
+          @click="closeDialog"
+        />
+        <v-spacer />
+      </v-card-actions>
+    </v-card>
+  </v-dialog>
+</template>
+
+<script setup lang="ts">
+import { isNil } from "lodash";
+import { computed, ref } from "vue";
+
+import AdministrationService from "@/api/service/AdministrationService";
+import { useSnackbarStore } from "@/store/SnackbarStore";
+import ReloadAuffaelligkeitenDTO from "@/types/config/ReloadAuffaelligkeitenDTO";
+import { useDateUtils } from "@/util/DateUtils";
+
+interface Props {
+  maxWidth: string;
+}
+defineProps<Props>();
+
+const snackbarStore = useSnackbarStore();
+const dateUtils = useDateUtils();
+
+const openDialogModel = ref(false);
+const dialogtext = ref("Hier könnte Ihre Werbung stehen.");
+
+const TOOLTIP_RELOAD_UNAUFFAELLIGER_TAG =
+  "Wenn es in Mobidam nachträgliche Änderungen bei den auffälligen Tagen gab, sollten diese neu geladen werden.";
+
+const dateToReload = ref(new Date());
+
+const choosenDate = computed({
+  get() {
+    let date = new Date();
+    if (!isNil(dateToReload.value)) {
+      date = dateToReload.value;
+      date.setHours(5);
+    }
+    return date;
+  },
+
+  set(date: Date) {
+    if (!isNil(date)) {
+      date.setHours(5);
+    }
+    dateToReload.value = date;
+  },
+});
+
+function closeDialog() {
+  dateToReload.value = new Date();
+  openDialogModel.value = false;
+}
+
+function openDialog(title: string) {
+  dialogtext.value = title;
+  openDialogModel.value = true;
+}
+
+function reloadUnauffaelligerTag() {
+  if (!isNil(dateToReload.value)) {
+    openDialog(
+      `Soll der ${dateUtils.getShortVersionOfDate(dateToReload.value)} erneut auf Auffälligkeiten überprüft werden?
+            </br> Die bereits ermittelten Auffälligkeiten werden dabei überschrieben.`
+    );
+  }
+}
+
+function confirmReload() {
+  if (dateToReload.value) {
+    const reloadAuffaelligkeiten = new ReloadAuffaelligkeitenDTO(
+      dateToReload.value
+    );
+    AdministrationService.reloadUnauffaelligerTag(reloadAuffaelligkeiten)
+      .then(() => {
+        snackbarStore.showSuccess(
+          `Der ${dateUtils.getShortVersionOfDate(reloadAuffaelligkeiten.dateToReload)} wurde erneut auf Auffälligkeiten überprüft.`
+        );
+      })
+      .catch((error) => {
+        snackbarStore.showError(error);
+      })
+      .finally(() => closeDialog());
+  }
+}
+</script>
