+++ frontend/src/types/config/ReloadAuffaelligkeitenDTO.ts
@@ -1,0 +1,7 @@
+export default class ReloadAuffaelligkeitenDTO {
+  dateToReload: Date;
+
+  constructor(dateToReload: Date) {
+    this.dateToReload = dateToReload;
+  }
+}
